[Refactor] #495 - viewModel 삭제 및 컨벤션에 맞게 blockUserViewCotoller 수정
